Move unit tests for RAL handlers.
