+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/text/distsql/ral/common/queryable/ShowAuthorityRuleHandlerTest.java
@@ -15,56 +15,51 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.text.distsql.ral.common.show;
+package org.apache.shardingsphere.proxy.backend.text.distsql.ral.common.queryable;
 
-import org.apache.shardingsphere.distsql.parser.statement.ral.common.show.ShowSQLParserRuleStatement;
+import org.apache.shardingsphere.authority.config.AuthorityRuleConfiguration;
+import org.apache.shardingsphere.distsql.parser.statement.ral.common.show.ShowAuthorityRuleStatement;
+import org.apache.shardingsphere.infra.config.algorithm.ShardingSphereAlgorithmConfiguration;
 import org.apache.shardingsphere.infra.metadata.rule.ShardingSphereRuleMetaData;
+import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
 import org.apache.shardingsphere.mode.manager.ContextManager;
-import org.apache.shardingsphere.parser.config.SQLParserRuleConfiguration;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
-import org.apache.shardingsphere.proxy.backend.text.distsql.ral.common.queryable.ShowSQLParserRuleHandler;
-import org.apache.shardingsphere.sql.parser.api.CacheOption;
 import org.junit.Test;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Properties;
 
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public final class ShowSQLParserRuleHandlerTest {
+public final class ShowAuthorityRuleHandlerTest {
+    
+    private final ShowAuthorityRuleHandler handler = new ShowAuthorityRuleHandler().initStatement(new ShowAuthorityRuleStatement());
     
     @Test
-    public void assertSQLParserRule() throws SQLException {
+    public void assertAuthorityRule() throws SQLException {
         ContextManager contextManager = mock(ContextManager.class, RETURNS_DEEP_STUBS);
         when(contextManager.getMetaDataContexts().getGlobalRuleMetaData()).thenReturn(getGlobalRuleMetaData());
         ProxyContext.getInstance().init(contextManager);
-        ShowSQLParserRuleHandler handler = new ShowSQLParserRuleHandler().initStatement(new ShowSQLParserRuleStatement());
         handler.execute();
         handler.next();
         List<Object> data = new ArrayList<>(handler.getRowData());
         assertThat(data.size(), is(3));
-        assertThat(data.get(0), is("true"));
-        String parseTreeCache = String.valueOf(data.get(1));
-        assertThat(parseTreeCache, containsString("\"initialCapacity\":128"));
-        assertThat(parseTreeCache, containsString("\"maximumSize\":1024"));
-        assertThat(parseTreeCache, containsString("\"concurrencyLevel\":4"));
-        String sqlStatementCache = String.valueOf(data.get(2));
-        assertThat(sqlStatementCache, containsString("\"initialCapacity\":2000"));
-        assertThat(sqlStatementCache, containsString("\"maximumSize\":65535"));
-        assertThat(sqlStatementCache, containsString("\"concurrencyLevel\":4"));
+        assertThat(data.get(0), is("root@localhost"));
+        assertThat(data.get(1), is("ALL_PRIVILEGES_PERMITTED"));
+        assertThat(data.get(2), is(""));
     }
     
     private ShardingSphereRuleMetaData getGlobalRuleMetaData() {
-        CacheOption parseTreeCache = new CacheOption(128, 1024, 4);
-        CacheOption sqlStatementCache = new CacheOption(2000, 65535, 4);
-        SQLParserRuleConfiguration sqlParserRuleConfiguration = new SQLParserRuleConfiguration(true, parseTreeCache, sqlStatementCache);
-        return new ShardingSphereRuleMetaData(Collections.singleton(sqlParserRuleConfiguration), Collections.emptyList());
+        ShardingSphereUser root = new ShardingSphereUser("root", "", "localhost");
+        AuthorityRuleConfiguration authorityRuleConfiguration = new AuthorityRuleConfiguration(Collections.singletonList(root),
+                new ShardingSphereAlgorithmConfiguration("ALL_PRIVILEGES_PERMITTED", new Properties()));
+        return new ShardingSphereRuleMetaData(Collections.singleton(authorityRuleConfiguration), Collections.emptyList());
     }
 }
